chore(gis): Refactor
