+++ modules/geopackage/src/lib/parse-geopackage.ts
@@ -1,6 +1,5 @@
 /* eslint-disable camelcase, @typescript-eslint/no-use-before-define */
 import {isBrowser} from '@loaders.gl/loader-utils';
-import {WKBLoader} from '@loaders.gl/wkt';
 import type {
   Schema,
   Field,
@@ -10,7 +9,7 @@ import type {
   GeoJSONTable,
   Feature
 } from '@loaders.gl/schema';
-import {binaryToGeometry, transformGeoJsonCoords} from '@loaders.gl/gis';
+import {convertWKBToGeoJSON, transformGeoJsonCoords} from '@loaders.gl/gis';
 import {Proj4Projection} from '@math.gl/proj4';
 import initSqlJs, {SqlJsStatic, Database, Statement} from 'sql.js';
 
@@ -399,12 +398,8 @@ function parseGeometry(arrayBuffer: ArrayBuffer): Geometry | null {
   // 2 byte magic, 1 byte version, 1 byte flags, 4 byte int32 srid
   const wkbOffset = 8 + envelopeLength;
 
-  // Loaders should not depend on `core` and the context passed to the main loader doesn't include a
-  // `parseSync` option, so instead we call parseSync directly on WKBLoader
-  const binaryGeometry = WKBLoader.parseSync?.(arrayBuffer.slice(wkbOffset));
-
-  // @ts-expect-error
-  return binaryToGeometry(binaryGeometry);
+  const wkbBytes = arrayBuffer.slice(wkbOffset);
+  return convertWKBToGeoJSON(wkbBytes);
 }
 
 /**
