+++ modules/gis/src/lib/feature-collection-converters/convert-binary-features-to-geojson.ts
@@ -3,28 +3,16 @@
 // Copyright (c) vis.gl contributors
 
 import type {
-  BinaryGeometry,
   BinaryGeometryType,
-  BinaryPointGeometry,
-  BinaryLineGeometry,
-  BinaryPolygonGeometry,
   BinaryFeatureCollection,
   BinaryFeature,
   // BinaryPointFeature,
   // BinaryLineFeature,
   // BinaryPolygonFeature,
-  BinaryAttribute,
   Feature,
-  Geometry,
-  Position,
-  GeoJsonProperties,
-  Point,
-  MultiPoint,
-  LineString,
-  MultiLineString,
-  Polygon,
-  MultiPolygon
+  GeoJsonProperties
 } from '@loaders.gl/schema';
+import {convertBinaryGeometryToGeoJSON} from '../geometry-converters/convert-binary-geometry-to-geojson';
 
 // Note:L We do not handle GeometryCollection, define a limited Geometry type that always has coordinates.
 // type FeatureGeometry = Point | MultiPoint | LineString | MultiLineString | Polygon | MultiPolygon;
@@ -42,7 +30,7 @@ type BinaryToGeoJsonOptions = {
  * @param options.featureId  Global feature id. If specified, only a single feature is extracted
  * @return GeoJSON objects
  */
-export function binaryToGeojson(
+export function convertBinaryFeatureCollectionToGeojson(
   data: BinaryFeatureCollection,
   options?: BinaryToGeoJsonOptions
 ): Feature[] | Feature {
@@ -89,25 +77,6 @@ function getSingleFeature(data: BinaryFeatureCollection, globalFeatureId: number
 function parseFeatures(data: BinaryFeatureCollection, type?: BinaryGeometryType): Feature[] {
   const dataArray = normalizeInput(data, type);
   return parseFeatureCollection(dataArray);
-}
-
-/** Parse input binary data and return a valid GeoJSON geometry object */
-export function binaryToGeometry(
-  data: BinaryGeometry,
-  startIndex?: number,
-  endIndex?: number
-): Geometry {
-  switch (data.type) {
-    case 'Point':
-      return pointToGeoJson(data, startIndex, endIndex);
-    case 'LineString':
-      return lineStringToGeoJson(data, startIndex, endIndex);
-    case 'Polygon':
-      return polygonToGeoJson(data, startIndex, endIndex);
-    default:
-      const unexpectedInput: never = data;
-      throw new Error(`Unsupported geometry type: ${(unexpectedInput as any)?.type}`);
-  }
 }
 
 // Normalize features
@@ -162,7 +131,7 @@ function parseFeatureCollection(dataArray: BinaryFeature[]): Feature[] {
 
 /** Parse input binary data and return a single GeoJSON Feature */
 function parseFeature(data: BinaryFeature, startIndex?: number, endIndex?: number): Feature {
-  const geometry = binaryToGeometry(data, startIndex, endIndex);
+  const geometry = convertBinaryGeometryToGeoJSON(data, startIndex, endIndex);
   const properties = parseProperties(data, startIndex, endIndex);
   const fields = parseFields(data, startIndex, endIndex);
   return {type: 'Feature', geometry, properties, ...fields};
@@ -180,110 +149,4 @@ function parseProperties(data, startIndex: number = 0, endIndex?: number): GeoJs
     properties[key] = data.numericProps[key].value[startIndex];
   }
   return properties;
-}
-
-/** Parse binary data of type Polygon */
-function polygonToGeoJson(
-  data: BinaryPolygonGeometry,
-  startIndex: number = -Infinity,
-  endIndex: number = Infinity
-): Polygon | MultiPolygon {
-  const {positions} = data;
-  const polygonIndices = data.polygonIndices.value.filter((x) => x >= startIndex && x <= endIndex);
-  const primitivePolygonIndices = data.primitivePolygonIndices.value.filter(
-    (x) => x >= startIndex && x <= endIndex
-  );
-  const multi = polygonIndices.length > 2;
-
-  // Polygon
-  if (!multi) {
-    const coordinates: Position[][] = [];
-    for (let i = 0; i < primitivePolygonIndices.length - 1; i++) {
-      const startRingIndex = primitivePolygonIndices[i];
-      const endRingIndex = primitivePolygonIndices[i + 1];
-      const ringCoordinates = ringToGeoJson(positions, startRingIndex, endRingIndex);
-      coordinates.push(ringCoordinates);
-    }
-
-    return {type: 'Polygon', coordinates};
-  }
-
-  // MultiPolygon
-  const coordinates: Position[][][] = [];
-  for (let i = 0; i < polygonIndices.length - 1; i++) {
-    const startPolygonIndex = polygonIndices[i];
-    const endPolygonIndex = polygonIndices[i + 1];
-    const polygonCoordinates = polygonToGeoJson(
-      data,
-      startPolygonIndex,
-      endPolygonIndex
-    ).coordinates;
-    coordinates.push(polygonCoordinates as Position[][]);
-  }
-
-  return {type: 'MultiPolygon', coordinates};
-}
-
-/** Parse binary data of type LineString */
-function lineStringToGeoJson(
-  data: BinaryLineGeometry,
-  startIndex: number = -Infinity,
-  endIndex: number = Infinity
-): LineString | MultiLineString {
-  const {positions} = data;
-  const pathIndices = data.pathIndices.value.filter((x) => x >= startIndex && x <= endIndex);
-  const multi = pathIndices.length > 2;
-
-  if (!multi) {
-    const coordinates = ringToGeoJson(positions, pathIndices[0], pathIndices[1]);
-    return {type: 'LineString', coordinates};
-  }
-
-  const coordinates: Position[][] = [];
-  for (let i = 0; i < pathIndices.length - 1; i++) {
-    const ringCoordinates = ringToGeoJson(positions, pathIndices[i], pathIndices[i + 1]);
-    coordinates.push(ringCoordinates);
-  }
-
-  return {type: 'MultiLineString', coordinates};
-}
-
-/** Parse binary data of type Point */
-function pointToGeoJson(data: BinaryPointGeometry, startIndex, endIndex): Point | MultiPoint {
-  const {positions} = data;
-  const coordinates = ringToGeoJson(positions, startIndex, endIndex);
-  const multi = coordinates.length > 1;
-
-  if (multi) {
-    return {type: 'MultiPoint', coordinates};
-  }
-
-  return {type: 'Point', coordinates: coordinates[0]};
-}
-
-/**
- * Parse a linear ring of positions to a GeoJSON linear ring
- *
- * @param positions Positions TypedArray
- * @param startIndex Start index to include in ring
- * @param endIndex End index to include in ring
- * @returns GeoJSON ring
- */
-function ringToGeoJson(
-  positions: BinaryAttribute,
-  startIndex?: number,
-  endIndex?: number
-): Position[] {
-  startIndex = startIndex || 0;
-  endIndex = endIndex || positions.value.length / positions.size;
-
-  const ringCoordinates: Position[] = [];
-  for (let j = startIndex; j < endIndex; j++) {
-    const coord = Array<number>();
-    for (let k = j * positions.size; k < (j + 1) * positions.size; k++) {
-      coord.push(Number(positions.value[k]));
-    }
-    ringCoordinates.push(coord);
-  }
-  return ringCoordinates;
 }
