+++ modules/gis/src/lib/feature-collection-converters/helpers/extract-geometry-info.ts
