+++ modules/gis/src/lib/feature-collection-converters/helpers/flat-geojson-to-binary-types.ts
@@ -1,3 +1,7 @@
+// loaders.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
 import type {TypedArray} from '@loaders.gl/schema';
 
 /**
