+++ modules/gis/src/lib/geoarrow/geoparquet-metadata-schema.ts
