+++ modules/gis/src/lib/geoarrow/geoparquet-metadata.ts
