+++ modules/gis/src/lib/geometry-converters/wkt/convert-wkb-to-binary-geometry.ts
@@ -14,9 +14,9 @@ import {
   concatenateBinaryPointGeometries,
   concatenateBinaryLineGeometries,
   concatenateBinaryPolygonGeometries
-} from '../../../api/binary-geometry-utils';
+} from '../../api/binary-geometry-utils';
 import {parseWKBHeader, WKBGeometryType} from './helpers/parse-wkb-header';
-import {concatTypedArrays} from '../../../utils/concat-typed-arrays';
+import {concatTypedArrays} from '../../utils/concat-typed-arrays';
 
 export function convertWKBToBinaryGeometry(arrayBuffer: ArrayBuffer): BinaryGeometry {
   const dataView = new DataView(arrayBuffer);
