+++ modules/gis/src/lib/geometry-converters/wkt/convert-wkb-to-geojson.ts
@@ -3,10 +3,10 @@
 // Copyright (c) vis.gl contributors
 
 import type {Geometry} from '@loaders.gl/schema';
-import {binaryToGeometry} from '../../../binary-features/binary-to-geojson';
+import {convertBinaryGeometryToGeoJSON} from '../convert-binary-geometry-to-geojson';
 import {convertWKBToBinaryGeometry} from './convert-wkb-to-binary-geometry';
 
 export function convertWKBToGeoJSON(arrayBuffer: ArrayBuffer): Geometry {
   const binaryGeometry = convertWKBToBinaryGeometry(arrayBuffer);
-  return binaryToGeometry(binaryGeometry);
+  return convertBinaryGeometryToGeoJSON(binaryGeometry);
 }
