+++ modules/gis/src/lib/geometry-converters/wkt/helpers/parse-wkb-header.ts
