+++ modules/gis/test/geoarrow/convert-geoarrow-to-geojson.spec.ts
@@ -11,10 +11,10 @@ import {
 import {load} from '@loaders.gl/core';
 import type {FeatureCollection} from '@loaders.gl/schema';
 import {convertArrowToSchema} from '@loaders.gl/schema-utils';
-import {getGeometryColumnsFromSchema, parseGeometryFromArrow} from '@loaders.gl/gis';
+import {getGeometryColumnsFromSchema, convertGeoArrowGeometryToGeoJSON} from '@loaders.gl/gis';
 import {ArrowLoader} from '@loaders.gl/arrow';
 
-test('ArrowUtils#parseGeometryFromArrow', async (t) => {
+test('ArrowUtils#convertGeoArrowGeometryToGeoJSON', async (t) => {
   for (const testCase of GEOARROW_TEST_CASES) {
     await testParseFromArrow(t, testCase[0], testCase[1]);
   }
@@ -68,7 +68,7 @@ async function testParseFromArrow(
     const firstArrowGeometry = table.getChild('geometry')?.get(0);
 
     // parse arrow geometry to geojson feature
-    const firstGeometry = parseGeometryFromArrow(firstArrowGeometry, encoding);
+    const firstGeometry = convertGeoArrowGeometryToGeoJSON(firstArrowGeometry, encoding);
 
     // check if geometry in firstFeature is equal to the original geometry in expectedPointGeojson
     t.deepEqual(
