+++ modules/schema/src/geometry/binary-geometries.ts
@@ -7,11 +7,38 @@ import type {TypedArray} from '../types/types';
 
 // BINARY FORMAT GEOMETRY
 
-export type BinaryAttribute = {value: TypedArray; size: number};
-export type BinaryGeometryType = 'Point' | 'LineString' | 'Polygon';
+/**
+ * Represent a collection of Features, similar to a GeoJSON FeatureCollection
+ * But in columnar format, with binary columns when possible
+ */
+export type BinaryFeatureCollection = {
+  shape: 'binary-feature-collection';
+  points?: BinaryPointFeature;
+  lines?: BinaryLineFeature;
+  polygons?: BinaryPolygonFeature;
+};
 
-type NumericProps = {[key: string]: BinaryAttribute};
-type Properties = object[];
+/** Binary feature + binary attributes */
+export type BinaryFeature = BinaryPointFeature | BinaryLineFeature | BinaryPolygonFeature;
+
+export type BinaryPointFeature = BinaryPointGeometry & BinaryProperties;
+export type BinaryLineFeature = BinaryLineGeometry & BinaryProperties;
+export type BinaryPolygonFeature = BinaryPolygonGeometry & BinaryProperties;
+
+/** Common properties for binary geometries */
+export type BinaryProperties = {
+  featureIds: BinaryAttribute;
+  globalFeatureIds: BinaryAttribute;
+  numericProps: Record<string, BinaryAttribute>;
+  properties: Record<string, any>[];
+  fields?: Record<string, any>[];
+};
+
+export type BinaryAttribute = {
+  value: TypedArray;
+  size: number;
+};
+export type BinaryGeometryType = 'Point' | 'LineString' | 'Polygon';
 
 /**
  * Represent a single Geometry, similar to a GeoJSON Geometry
@@ -38,30 +65,4 @@ export type BinaryPolygonGeometry = {
   polygonIndices: BinaryAttribute;
   primitivePolygonIndices: BinaryAttribute;
   triangles?: BinaryAttribute;
-};
-
-/** Common properties for binary geometries */
-export type BinaryProperties = {
-  featureIds: BinaryAttribute;
-  globalFeatureIds: BinaryAttribute;
-  numericProps: NumericProps;
-  properties: Properties;
-  fields?: Properties;
-};
-
-/** Binary feature + binary attributes */
-export type BinaryFeature = BinaryPointFeature | BinaryLineFeature | BinaryPolygonFeature;
-
-export type BinaryPointFeature = BinaryPointGeometry & BinaryProperties;
-export type BinaryLineFeature = BinaryLineGeometry & BinaryProperties;
-export type BinaryPolygonFeature = BinaryPolygonGeometry & BinaryProperties;
-
-/**
- * Represent a collection of Features, similar to a GeoJSON FeatureCollection
- */
-export type BinaryFeatureCollection = {
-  shape: 'binary-feature-collection';
-  points?: BinaryPointFeature;
-  lines?: BinaryLineFeature;
-  polygons?: BinaryPolygonFeature;
 };
